math -> cmath

+++ libgg/Vec2.h
@@ -33,7 +33,7 @@ public:
 			m_v[0] * m_v[0] +
 			m_v[1] * m_v[1];
 		
-		return std::sqrt(sq);
+		return std::sqrtf(sq);
 	}
 
 	inline float CalcSizeSq() const
