+++ libgg/Vec3.h
@@ -29,7 +29,7 @@ public:
 			m_v[1] * m_v[1] +
 			m_v[2] * m_v[2];
 		
-		return std::sqrt(sq);
+		return std::sqrtf(sq);
 	}
 
 	inline float CalcSizeSq() const
